move usage and versions to info file
